+++ src/FluentDbTools/Abstractions/FluentDbTools.Migration.Abstractions/SqlStatement.cs
@@ -1,0 +1,8 @@
+﻿namespace FluentDbTools.Migration.Abstractions
+{
+    internal class SqlStatement
+    {
+        public string Sql { get; set; }
+        public bool IsExternal { get; set; }
+    }
+}