+++ src/FluentDbTools/Extensions/FluentDbTools.Extensions.Migration/ConfigureWithMigrationExtensions.cs
@@ -1,9 +1,15 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using FluentDbTools.Common.Abstractions;
 using FluentDbTools.Extensions.MSDependencyInjection;
 using FluentDbTools.Extensions.Migration.DefaultConfigs;
 using FluentDbTools.Migration.Abstractions;
+using FluentDbTools.Migration.Oracle;
+using FluentDbTools.Migration.Postgres;
+using FluentMigrator;
+using FluentMigrator.Runner.Generators.Oracle;
+using FluentMigrator.Runner.Generators.Postgres;
 using FluentMigrator.Runner.VersionTableInfo;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
@@ -13,7 +19,7 @@ namespace FluentDbTools.Extensions.Migration
     public static class ConfigureWithMigrationExtensions
     {
         public static IServiceCollection ConfigureWithMigrationAndScanForVersionTable(
-            this IServiceCollection serviceCollection, 
+            this IServiceCollection serviceCollection,
             IEnumerable<Assembly> assembliesWithMigrationModels)
         {
             var assembliesWithMigrationModelArray = assembliesWithMigrationModels.ToArray();
@@ -24,16 +30,16 @@ namespace FluentDbTools.Extensions.Migration
         }
 
         public static IServiceCollection ConfigureWithMigration(
-            this IServiceCollection serviceCollection, 
+            this IServiceCollection serviceCollection,
             IEnumerable<Assembly> assembliesWithMigrationModels)
         {
             var assembliesWithMigrationModelArray = assembliesWithMigrationModels.ToArray();
             return serviceCollection
+                .AddDefaultDbMigrationConfig()
                 .Register(FluentDbTools.Migration.Oracle.ServiceRegistration.Register)
                 .Register(FluentDbTools.Migration.Postgres.ServiceRegistration.Register)
-                .AddDefaultDbMigrationConfig()
-                .ConfigureWithMigrationAssemblies(FluentDbTools.Migration.ServiceRegistration.Register,
-                    assembliesWithMigrationModelArray);
+                .RegisterDatabaseDependedFluentMigrationTypes()
+                .ConfigureWithMigrationAssemblies(FluentDbTools.Migration.ServiceRegistration.Register, assembliesWithMigrationModelArray);
         }
 
         internal static IServiceCollection ConfigureVersionTableMetaData(
@@ -66,6 +72,32 @@ namespace FluentDbTools.Extensions.Migration
             return serviceCollection;
         }
 
+        public static IServiceCollection RegisterDatabaseDependedFluentMigrationTypes(this IServiceCollection serviceCollection)
+        {
+            return serviceCollection
+            .AddScoped<IMigrationProcessor>(sp =>
+            {
+                var isPostgres = sp.GetService<IDbMigrationConfig>()?.DbType == SupportedDatabaseTypes.Postgres;
+                if (isPostgres)
+                {
+                    return sp.GetRequiredService<ExtendedPostgresProcessor>();
+                }
+
+                return sp.GetRequiredService<ExtendedOracleManagedProcessor>();
+            })
+            .AddScoped<IMigrationGenerator>(sp =>
+            {
+                var isPostgres = sp.GetService<IDbMigrationConfig>()?.DbType == SupportedDatabaseTypes.Postgres;
+                if (isPostgres)
+                {
+                    return sp.GetRequiredService<PostgresGenerator>();
+                }
+
+                return sp.GetRequiredService<OracleGenerator>();
+            });
+
+
+        }
 
     }
 }