+++ src/FluentDbTools/Implementations/FluentDbTools.Migration/FluentMigrationExtensions.cs
@@ -1,4 +1,5 @@
-﻿using FluentMigrator.Expressions;
+﻿using FluentDbTools.Migration.Abstractions;
+using FluentMigrator.Expressions;
 using FluentMigrator.Runner;
 using FluentMigrator.Runner.VersionTableInfo;
 
@@ -9,8 +10,15 @@ namespace FluentDbTools.Migration
         public static void DropSchema(this IMigrationRunner migrationRunner, IVersionTableMetaData versionTableMetaData)
         {
             var schemaName = versionTableMetaData.SchemaName;
-            migrationRunner.Processor
-                .Process(new DeleteSchemaExpression { SchemaName = schemaName });
+            migrationRunner.Processor.Process(new DeleteSchemaExpression { SchemaName = schemaName });
         }
+
+
+        public static void CreateSchema(this IMigrationRunner migrationRunner, IVersionTableMetaData versionTableMetaData)
+        {
+            var schemaName = versionTableMetaData.SchemaName;
+            migrationRunner.Processor.Process(new CreateSchemaExpression { SchemaName = schemaName });
+        }
+
     }
 }